requirements.txt
streamlit==1.31.0
pandas==2.2.0
numpy==1.26.3
plotly==5.18.0
sqlalchemy==2.0.25
faker==22.6.0
python-dotenv==1.0.1
scikit-learn==1.4.0
textblob==0.17.1
openai==1.12.0
anthropic==0.18.0
fastapi==0.109.2
uvicorn==0.27.1
